+++ src/constants/editioncard.ts
@@ -5,3 +5,4 @@ export const MAX_PUBLISHER_NAME_LENGTH = 80;
 export const MAX_PLACE_LENGTH = 80;
 export const PLACEHOLDER_COVER_LINK =
   "https://test-sfr-covers.s3.amazonaws.com/default/defaultCover.png";
+export const MAX_PAGE_TITLE_LENGTH = 45;