+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Feedback button test PW
 - SFR-1797: Remove Playwright tests with clicks on header and footer links
 - Update README with C4 diagrams and epub-to-webpub info
+- Update Collection, Edition, and Work page titles for accessibility
 
 ## [0.17.4]